change the limit= meaning
now the `limit=0` means "no any records" (was "no limit")
and `limit=-1` means "no limit"

so the `/search?limit=0` means the same as `/count`
except the output format.

+++ search/ryftprim/task.go
@@ -107,10 +107,10 @@ func (task *Task) startProcessing(engine *Engine, res *search.Result) {
 		task.ViewFileName, task.config.Delimiter)
 
 	// result reader options
-	rr.Offset = uint64(task.config.Offset) // start from this record
-	rr.Limit = uint64(task.config.Limit)   // limit the total number of records
-	rr.ReadData = task.config.ReportData   // if `false` only indexes will be reported
-	rr.MakeView = !task.isShow             // if /show do not create VIEW file, just use it
+	rr.Offset = task.config.Offset       // start from this record
+	rr.Limit = task.config.Limit         // limit the total number of records
+	rr.ReadData = task.config.ReportData // if `false` only indexes will be reported
+	rr.MakeView = !task.isShow           // if /show do not create VIEW file, just use it
 
 	// report filepath relative to home and update index's host
 	rr.RelativeToHome = filepath.Join(engine.MountPoint, engine.HomeDir)
